Refactor app screens into separate files.
