+++ bitvault-ui/src/app/screens.rs
@@ -1,0 +1,6 @@
+// Re-export screen modules
+pub mod home;
+pub mod lock;
+pub mod onboarding;
+pub mod seed;
+pub mod wallet;
